thestudio-production-test-rig/pyproject.toml
[project]
name = "thestudio-production-test-rig"
version = "0.1.0"
description = "Dedicated test suite for deployed TheStudio (production/staging)"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=6.0.0",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
